Separate the game logic from the sector map
There was a need to use the sector map in different contexts, not only
inside the SectorView. Therefore, the 3d star map viewer has been moved
to a separate SectorMap class. The SectorMap class has an interface for
adding custom objects to the map - lines, spheres and round indicators.
The SectorView uses this to overlay game logic objects (player location,
route, etc.) on the map. Some functions that SectorView used to perform
directly are passed to SectorMap, and are performed via
GetMap()->Function(). The SectorMapContext class has also been added,
which contains all the necessary external information for the SectorMap
to function.

Also added the ability to display SectorMap inside the ImGui window.

Some additional modifications:

  * completely get rid of Pi:: in the SectorMap class
    - move GetMoveSpeedShiftModifier() to Input::Manager
    - explicitly pass frame time to SectorMap::Update
  * input tweaking
    - receive wheel data directly instead of via a signal
