+++ src/Input.h
@@ -220,6 +220,10 @@ public:
 	void SetCapturingMouse(bool enabled);
 	void ClearMouse();
 
+	// Get the default speed modifier to apply to movement (scrolling, zooming...), depending on the "shift" keys.
+	// This is a default value only, centralized here to promote uniform user expericience.
+	float GetMoveSpeedShiftModifier();
+
 	sigc::signal<void, SDL_Keysym *> onKeyPress;
 	sigc::signal<void, SDL_Keysym *> onKeyRelease;
 	sigc::signal<void, int, int, int> onMouseButtonUp;
